requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
marshmallow
flask-marshmallow
flask-restful

